ref: rename parse mod to syntax mod
